+++ pySDC/tests/test_projects/test_resilience/test_Lorenz.py
@@ -3,6 +3,6 @@ import pytest
 
 @pytest.mark.base
 def test_main():
-    from pySDC.projects.Resilience.Lorentz import main
+    from pySDC.projects.Resilience.Lorenz import main
 
     main(plotting=False)
